Backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
bson==0.5.10
click==8.1.7
dnspython==2.6.1
fastapi==0.115.0
h11==0.14.0
idna==3.10
motor==3.6.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.1
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlmodel==0.0.22
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.0
